Split absolute_url helper into absolute_url_for_site and absolute_url_for_request.
There are two ways of building an absolute URL in the app: looking at the domain
of the Site record in the database, and looking at the one being accessed by the
current user.

The former often makes most sense to the user, especially if using a
non-standard internal URL or a site that hasn't been fully configured,
but it's not possible when generating an email message outside of a
request context for example, so you need the latter in that case.

Depending on how the URLs were generated by the app, you may need the
test to generate a URL in the same kind of way, that will match, to
prove that the app did it correctly. Before, the way of doing this was
ambiguous, but now it's clear. This breaks backwards compatibility with
tests using the old, ambiguous way.